Fix npm Issue, Use sdkmanager, Add Emulator, Install Cordova and Ionic
Fix npm Issue:
npm clean cache needs the force argument to work.

Use sdkmanager:
This approach requires Android SDK tools 26.
We install latest Android SDK tools, latest platforms, and latest build
tools.

Add Emulator - it could be useful for testing.
Users can use this image as a base and then install their own android
system image on top of it.

Use Ubuntu 17.10 Image - this brings latest stable versions of packages
and could fix some bugs with them.

Silence Commands - the logs are huge so we redirect stdout to /dev/null.
Warnings and errors appear on stderr so we do not need usually care for
stdout.

Add Latest x86_64 System Image
Since we included an emulator, we can include one system image too. The
images are quite large so we install only the latest SDK. Users can use
the docker image as a base and install their own android system image on
top of it. We install the one with Google APIs. The latest version of
this image is 25.

Install Cordova and Ionic - they can be useful for some apps.