CI: Build using docker action
Change to build the image using docker action, this should then allow
the docker action's cache to be used. Subsequently reducing build time
~50%, as the image will only need to be built once. Currently image is
built twice.

The default driver uses double the disk space, see
docker/build-push-action/issues/321 (in brief the image is build in the
build-push-action local cache and then transfered to the local docker).
This is a problem as this image is so large. Using the `docker` driver
will workaround this.