Setup jEnv branching
The JENV_TAGGED build argument was added with accepted values of
"tagged" or "latest" (default), quotes excluded. When using the
"tagged" value this allows for a specific tag of jEnv to be used, with a
default set. The JENV_VER argument allows one to set the tag of jEnv to
be installed when building.
The default of "latest" uses the master branch of jEnv to be installed.