Modify how Bundletool Version is grabbed
Previously the bundletool version was not being grabbed correctly when
run resulting in an empty Bundletool Version string within the final
INSTALLED_VERSIONS. This may have been a result of way the characters
`tr` was deleting was written, they where escaped. As such it is
believed that the escape needed to be double escaped.
The change made was to throw the characters `tr` was looking for into
quotes. This still results in a correctly grabbed Bundletool Version
when testing locally.