PR #179: CI: Fixup build and push #178 & minor improvements @master-bob
When tags where build & pushed to Docker Hub, the build arguments added
in as a result of
https://github.com/mingchen/docker-android-build-box/issues/116 where
not included. This fix should make the tagged version use the correct
build arguments. By doing so the correct docker image should be built
and pushed to Docker Hub.

Tags which are affected by this on Docker Hub which should be re-pushed
are `1.27.0` and `1.26.0`.

This should close #178 .